Remove emitter from class Repair
